Revert "fix issue with re-encrypting"
This reverts commit 9b265bc0fe3530d2659439ac5c47817d40559f4f.

The previous behavior was correct. helm-secrets should change the working
directory, since sops searches from the working directory and upwards for
the sops config file to use. Not doing so can cause unpredictable behavior,
where different (or no) config files are found and used depending on the
current working directory.